Remove old options dialog
