+++ PlattformOrdMan/UI/Dialog/OptionsDialog/FilteringOptionsDialog.resx
