--- PlattformOrdMan/UI/Dialog/OptionsDialog/OrderHistoryOptionOK.cs
@@ -1,4 +1,0 @@
-﻿namespace PlattformOrdMan.UI.Dialog.OptionsDialog
-{
-    public delegate void OrderHistoryOptionOK(bool isIncludedColumnsUpdated);
-}
